IdentityCheckerを実装
